Make invertibility methods available outside CheckBoundsDeclarations (#1100)
* Move LValueOccurrenceCount and VariableOccurrenceCount into ExprUtil class

* Move SplitByLValueCount into ExprUtil class

* Move ReadsMemoryViaPointer to ExprUtil class

* Move CheckIsNonModifying method to the ExprUtil class

* Move EqualValue into the ExprUtil class

* Move IsInvertible methods into a new InverseUtil class

* Move IgnoreRedundantCasts from BoundsUtil into ExprUtil

* Move CreateExplicitCast into ExprCreatorUtil

* Move Inverse methods into InverseUtil
