+++ clang/lib/Sema/SemaBounds.cpp
@@ -68,19 +68,6 @@ public:
       K == BoundsExpr::Kind::Range || K == BoundsExpr::Kind::Invalid);
   }
 
-  static Expr *IgnoreRedundantCast(ASTContext &Ctx, CastKind NewCK, Expr *E) {
-    CastExpr *P = dyn_cast<CastExpr>(E);
-    if (!P)
-      return E;
-
-    CastKind ExistingCK = P->getCastKind();
-    Expr *SE = P->getSubExpr();
-    if (NewCK == CK_BitCast && ExistingCK == CK_BitCast)
-      return SE;
-
-    return E;
-  }
-
   static bool getReferentSizeInChars(ASTContext &Ctx, QualType Ty, llvm::APSInt &Size) {
     assert(Ty->isPointerType());
     const Type *Pointee = Ty->getPointeeOrArrayElementType();
@@ -477,91 +464,6 @@ namespace {
     else
       return R.get();
   }	
-}
-
-namespace {
-  class LValueCountHelper : public RecursiveASTVisitor<LValueCountHelper> {
-    private:
-      Sema &SemaRef;
-      Lexicographic Lex;
-      Expr *LValue;
-      ValueDecl *V;
-      unsigned int Count;
-
-    public:
-      LValueCountHelper(Sema &SemaRef, Expr *LValue, ValueDecl *V) :
-        SemaRef(SemaRef),
-        Lex(Lexicographic(SemaRef.Context, nullptr)),
-        LValue(LValue),
-        V(V),
-        Count(0) {}
-
-      unsigned int GetCount() { return Count; }
-
-      bool VisitDeclRefExpr(DeclRefExpr *E) {
-        // Check for an occurrence of a variable whose declaration matches V.
-        if (V) {
-          if (ValueDecl *D = E->getDecl()) {
-            if (Lex.CompareDecl(D, V) == Lexicographic::Result::Equal)
-              ++Count;
-          }
-          return true;
-        }
-
-        // Check for an occurrence of a variable equal to LValue if LValue
-        // is a variable.
-        DeclRefExpr *Var = dyn_cast_or_null<DeclRefExpr>(LValue);
-        if (!Var)
-          return true;
-        if (Lex.CompareExpr(Var, E) == Lexicographic::Result::Equal)
-          ++Count;
-        return true;
-      }
-
-      bool VisitMemberExpr(MemberExpr *E) {
-        MemberExpr *M = dyn_cast_or_null<MemberExpr>(LValue);
-        if (!M)
-          return true;
-        if (Lex.CompareExprSemantically(E, M))
-          ++Count;
-        return true;
-      }
-
-      // Do not traverse the child of a BoundsValueExpr.
-      // If a BoundsValueExpr uses the expression LValue (or a variable whose
-      // declaration matches V), this should not count toward the total
-      // occurrence count of LValue or V in the expression.
-      // For example, for the expression BoundsValue(TempBinding(v)) + v, the
-      // total occurrence count of the variable v should be 1, not 2.
-      bool TraverseBoundsValueExpr(BoundsValueExpr *E) {
-        return true;
-      }
-  };
-
-  // VariableOccurrenceCount returns the number of occurrences of variable
-  // expressions in E whose Decls are equivalent to V.
-  unsigned int VariableOccurrenceCount(Sema &SemaRef, ValueDecl *V, Expr *E) {
-    if (!V)
-      return 0;
-    LValueCountHelper Counter(SemaRef, nullptr, V);
-    Counter.TraverseStmt(E);
-    return Counter.GetCount();
-  }
-
-  // VariableOccurrenceCount returns the number of occurrences of the Target
-  // variable expression in E.
-  unsigned int VariableOccurrenceCount(Sema &SemaRef, DeclRefExpr *Target,
-                                       Expr *E) {
-    return VariableOccurrenceCount(SemaRef, Target->getDecl(), E);
-  }
-
-  // LValueOccurrenceCount returns the number of occurrences of the LValue
-  // expression in E.
-  unsigned int LValueOccurrenceCount(Sema &SemaRef, Expr *LValue, Expr *E) {
-    LValueCountHelper Counter(SemaRef, LValue, nullptr);
-    Counter.TraverseStmt(E);
-    return Counter.GetCount();
-  }
 }
 
 namespace {
@@ -1549,7 +1451,7 @@ namespace {
           return ProofResult::Maybe;
         }
 
-        if (EqualValue(S.Context, Base, R.Base, EquivExprs)) {
+        if (ExprUtil::EqualValue(S.Context, Base, R.Base, EquivExprs)) {
           ProofResult LowerBoundsResult = CompareLowerOffsetsImpl(R, Cause, EquivExprs, Facts);
           ProofResult UpperBoundsResult = CompareUpperOffsetsImpl(R, Cause, EquivExprs, Facts);
 
@@ -1605,7 +1507,7 @@ namespace {
         FreeVariablePosition ObservedBasePos = CombineFreeVariablePosition(
             FreeVariablePosition::Observed, BasePos);
 
-        if (EqualValue(S.Context, Base, R.Base, EquivExprs)) {
+        if (ExprUtil::EqualValue(S.Context, Base, R.Base, EquivExprs)) {
           ProofResult LowerBoundsResult =
               CompareLowerOffsets(R, Cause, EquivExprs, Facts, FreeVariables);
           ProofResult UpperBoundsResult =
@@ -1774,7 +1676,8 @@ namespace {
         // We also skip checking free variables if E1 or E2 is or contains a
         // non-arrow member expression, since the compiler currently does
         // not track equality information for member expressions.
-        if (ReadsMemoryViaPointer(E1, true) || ReadsMemoryViaPointer(E2, true))
+        if (ExprUtil::ReadsMemoryViaPointer(E1, true) ||
+            ExprUtil::ReadsMemoryViaPointer(E2, true))
           return false;
 
         bool HasFreeVariables = false;
@@ -1835,7 +1738,7 @@ namespace {
             continue;
           auto It = DstVars.begin();
           for (; It != DstVars.end(); It++) {
-            if (EqualValue(S.Context, SrcV, *It, EquivExprs))
+            if (ExprUtil::EqualValue(S.Context, SrcV, *It, EquivExprs))
               break;
           }
 
@@ -1875,7 +1778,7 @@ namespace {
           // Search InnerList for an expression that uses the value of SrcV.
           for (; SrcIndex < InnerListSize; ++SrcIndex) {
             Expr *E = InnerList[SrcIndex];
-            SrcVarCount = VariableOccurrenceCount(S, SrcV, E);
+            SrcVarCount = ExprUtil::VariableOccurrenceCount(S, SrcV, E);
             if (SrcVarCount > 0)
               break;
           }
@@ -1893,7 +1796,7 @@ namespace {
             for (auto I = DstVars.begin(); I != DstVars.end(); ++I) {
               DeclRefExpr *DstV = cast<DeclRefExpr>(*I);
               Expr *E = InnerList[DstIndex];
-              if (VariableOccurrenceCount(S, DstV, E) > 0)
+              if (ExprUtil::VariableOccurrenceCount(S, DstV, E) > 0)
                 return true;
             }
           }
@@ -2221,7 +2124,7 @@ namespace {
       if (!Offset1 && !Offset2)
         return false;
 
-      if (!EqualValue(Ctx, Base1, Base2, EquivExprs))
+      if (!ExprUtil::EqualValue(Ctx, Base1, Base2, EquivExprs))
         return false;
 
       llvm::APSInt ConstantPart1, ConstantPart2;
@@ -2233,7 +2136,7 @@ namespace {
       
       if (!CreatedStdForm1 || !CreatedStdForm2)
         return false;
-      if (!EqualValue(Ctx, VariablePart1, VariablePart2, EquivExprs))
+      if (!ExprUtil::EqualValue(Ctx, VariablePart1, VariablePart2, EquivExprs))
         return false;
       if (IsOpSigned1 != IsOpSigned2)
         return false;
@@ -2254,7 +2157,7 @@ namespace {
       if (!Offset1 && !Offset2)
         return false;
 
-      if (!EqualValue(Ctx, Base1, Base2, EquivExprs))
+      if (!ExprUtil::EqualValue(Ctx, Base1, Base2, EquivExprs))
         return false;
 
       llvm::APSInt ConstantPart1, ConstantPart2;
@@ -2271,7 +2174,7 @@ namespace {
       if (ConstantPart1 != ConstantPart2)
         return false;
 
-      if (EqualValue(Ctx, VariablePart1, VariablePart2, EquivExprs))
+      if (ExprUtil::EqualValue(Ctx, VariablePart1, VariablePart2, EquivExprs))
         return true;
       if (FactExists(Ctx, VariablePart1, VariablePart2, EquivExprs, Facts))
         return true;
@@ -2299,11 +2202,6 @@ namespace {
         }
       }
       return ExistsIn && !ExistsKill;
-    }
-
-    static bool EqualValue(ASTContext &Ctx, Expr *E1, Expr *E2, EquivExprSets *EquivExprs) {
-      Lexicographic::Result R = Lexicographic(Ctx, EquivExprs).CompareExpr(E1, E2);
-      return R == Lexicographic::Result::Equal;
     }
 
     // Convert the bounds expression `Bounds` to a range `R`. This function returns true
@@ -2342,7 +2240,7 @@ namespace {
 
           // If both of the offsets are constants, the range is considered constant-sized.
           // Otherwise, it is a variable-sized range.
-          if (EqualValue(S.Context, LowerBase, UpperBase, EquivExprs)) {
+          if (ExprUtil::EqualValue(S.Context, LowerBase, UpperBase, EquivExprs)) {
             R->SetBase(LowerBase);
             R->SetLowerConstant(LowerOffsetConstant);
             R->SetLowerVariable(LowerOffsetVariable);
@@ -3091,7 +2989,7 @@ namespace {
        for (auto InnerList = (*I).begin(); InnerList != (*I).end();
                                                          ++InnerList) {
          Expr *E = *InnerList;
-         if (!VariableOccurrenceCount(S, V, E))
+         if (!ExprUtil::VariableOccurrenceCount(S, V, E))
            ExprList.push_back(E);
        }
        if (ExprList.size() > 1)
@@ -3850,8 +3748,8 @@ namespace {
             // The bounds expression is for an interface type. Retype the
             // argument to the interface type.
             if (UsedIType) {
-              TypedArg = CreateExplicitCast(
-                ParamIType->getType(), CK_BitCast, Arg, true);
+              TypedArg = ExprCreatorUtil::CreateExplicitCast(
+                S, ParamIType->getType(), CK_BitCast, Arg, true);
             }
             SubstParamBounds = ExpandToRange(TypedArg,
                                     const_cast<BoundsExpr *>(SubstParamBounds));
@@ -3913,12 +3811,12 @@ namespace {
       // same value as e.
       if (CK == CastKind::CK_ArrayToPointerDecay) {
         // State.SameValue = { e } for lvalues with array type.
-        if (!CreatesNewObject(E) && CheckIsNonModifying(E))
+        if (!CreatesNewObject(E) && ExprUtil::CheckIsNonModifying(S, E))
           State.SameValue = { E };
       } else if (CK == CastKind::CK_LValueToRValue) {
         if (E->getType()->isArrayType()) {
-          // State.Same = { e } for lvalues with array type.
-          if (!CreatesNewObject(E) && CheckIsNonModifying(E))
+          // State.SameValue = { e } for lvalues with array type.
+          if (!CreatesNewObject(E) && ExprUtil::CheckIsNonModifying(S, E))
             State.SameValue = { E };
         }
         else {
@@ -3929,7 +3827,8 @@ namespace {
             // Otherwise, if e is nonmodifying and does not read memory via a
             // pointer, State.SameValue = { e }.  Otherwise, State.SameValue
             // is empty.
-            if (CheckIsNonModifying(E) && !ReadsMemoryViaPointer(E) &&
+            if (ExprUtil::CheckIsNonModifying(S, E) &&
+                !ExprUtil::ReadsMemoryViaPointer(E) &&
                 !CreatesNewObject(E))
               State.SameValue.push_back(E);
           }
@@ -3978,9 +3877,10 @@ namespace {
         // recompute any expressions computed to temporaries already.
         Expr *TempUse = CreateTemporaryUse(TempExpr);
 
-        Expr *SubExprAtNewType = CreateExplicitCast(E->getType(),
-                                                CastKind::CK_BitCast,
-                                                TempUse, true);
+        Expr *SubExprAtNewType =
+          ExprCreatorUtil::CreateExplicitCast(S, E->getType(),
+                                              CastKind::CK_BitCast,
+                                              TempUse, true);
 
         if (CK == CK_AssumePtrBounds)
           return ExpandToRange(SubExprAtNewType, E->getBoundsExpr());
@@ -4383,7 +4283,7 @@ namespace {
 
       State.SameValue = IntersectExprSets(StateTrueArm.SameValue,
                                           StateFalseArm.SameValue);
-      if (!CreatesNewObject(E) && CheckIsNonModifying(E) &&
+      if (!CreatesNewObject(E) && ExprUtil::CheckIsNonModifying(S, E) &&
           !EqualExprsContainsExpr(State.SameValue, E))
         State.SameValue.push_back(E);
 
@@ -4632,20 +4532,6 @@ namespace {
       return CBE;
     }
 
-    Expr *CreateExplicitCast(QualType Target, CastKind CK, Expr *E,
-                               bool isBoundsSafeInterface) {
-      // Avoid building up nested chains of no-op casts.
-      E = BoundsUtil::IgnoreRedundantCast(Context, CK, E);
-
-      // Synthesize some dummy type source source information.
-      TypeSourceInfo *DI = Context.getTrivialTypeSourceInfo(Target);
-      CStyleCastExpr *CE = CStyleCastExpr::Create(Context, Target,
-        ExprValueKind::VK_RValue, CK, E, nullptr, DI, SourceLocation(),
-        SourceLocation());
-      CE->setBoundsSafeInterface(isBoundsSafeInterface);
-      return CE;
-    }
-
     ImplicitCastExpr *CreateImplicitCast(QualType Target, CastKind CK,
                                          Expr *E) {
       return ImplicitCastExpr::Create(Context, Target, CK, E, nullptr,
@@ -4684,7 +4570,9 @@ namespace {
             // The bounds-safe interface argument is false because casts
             // to checked pointer types are always allowed by type checking.
             LowerBound =
-              CreateExplicitCast(ResultTy, CastKind::CK_BitCast, Base, false);
+              ExprCreatorUtil::CreateExplicitCast(S, ResultTy,
+                                                  CastKind::CK_BitCast,
+                                                  Base, false);
           } else {
             ResultTy = Base->getType();
             LowerBound = Base;
@@ -4695,7 +4583,9 @@ namespace {
               // between checked pointer types are always allowed by type
               // checking.
               LowerBound =
-                CreateExplicitCast(ResultTy, CastKind::CK_BitCast, Base, false);
+                ExprCreatorUtil::CreateExplicitCast(S, ResultTy,
+                                                    CastKind::CK_BitCast,
+                                                    Base, false);
             }
           }
           Expr *UpperBound =
@@ -5252,7 +5142,7 @@ namespace {
         CHKCBindTemporaryExpr *Temp = GetTempBinding(Src);
         if (Temp)
           State.TargetSrcEquality[Target] = CreateTemporaryUse(Temp);
-        else if (CheckIsNonModifying(Src))
+        else if (ExprUtil::CheckIsNonModifying(S, Src))
           State.TargetSrcEquality[Target] = Src;
       }
 
@@ -5344,7 +5234,7 @@ namespace {
       }
 
       // If Val is a non-modifying expression, SameValue contains Val.
-      else if (CheckIsNonModifying(Val))
+      else if (ExprUtil::CheckIsNonModifying(S, Val))
         SameValue.push_back(Val);
 
       // If Val is a modifying expression, use the SameValue_i sets of
@@ -5358,11 +5248,11 @@ namespace {
           // For any modifying subexpression SubExpr_i of e, try to set
           // ValPrime to a nonmodifying expression from the set SameValue_i
           // of expressions that produce the same value as SubExpr_i.
-          if (!CheckIsNonModifying(SubExpr_i)) {
+          if (!ExprUtil::CheckIsNonModifying(S, SubExpr_i)) {
             ExprSetTy SameValue_i = Pair.second;
             for (auto I = SameValue_i.begin(); I != SameValue_i.end(); ++I) {
               Expr *E_i = *I;
-              if (CheckIsNonModifying(E_i)) {
+              if (ExprUtil::CheckIsNonModifying(S, E_i)) {
                 ValPrime = E_i;
                 break;
               }
@@ -5402,8 +5292,8 @@ namespace {
                            bool &OriginalValueUsesLValue) {
       // Check if Src has an inverse expression with respect to LValue.
       Expr *IV = nullptr;
-      if (IsInvertible(LValue, Src))
-        IV = Inverse(LValue, Target, Src);
+      if (InverseUtil::IsInvertible(S, LValue, Src))
+        IV = InverseUtil::Inverse(S, LValue, Target, Src);
       if (IV) {
         // If Src has an inverse with respect to LValue, then the original
         // value (the inverse) must use the value of LValue.
@@ -5436,7 +5326,7 @@ namespace {
         Lexicographic Lex(S.Context, nullptr);
         Expr *E = Lex.IgnoreValuePreservingOperations(S.Context, *I);
         DeclRefExpr *W = VariableUtil::GetRValueVariable(S, E);
-        if (W != nullptr && !EqualValue(S.Context, V, W, nullptr)) {
+        if (W != nullptr && !ExprUtil::EqualValue(S.Context, V, W, nullptr)) {
           // Expression equality in EquivExprs does not account for types, so
           // expressions in the same set in EquivExprs may not have the same
           // type. The original value of Src with respect to v must have a type
@@ -5451,306 +5341,6 @@ namespace {
         }
       }
 
-      return nullptr;
-    }
-
-    bool IsInvertible(Expr *LValue, Expr *E) {
-      if (!E)
-        return false;
-
-      E = E->IgnoreParens();
-      Expr *RValueChild = ExprUtil::GetRValueCastChild(S, E);
-      if (RValueChild && EqualValue(S.Context, LValue, RValueChild, nullptr))
-        return true;
-
-      switch (E->getStmtClass()) {
-        case Expr::UnaryOperatorClass:
-          return IsUnaryOperatorInvertible(LValue, cast<UnaryOperator>(E));
-        case Expr::BinaryOperatorClass:
-          return IsBinaryOperatorInvertible(LValue, cast<BinaryOperator>(E));
-        case Expr::ImplicitCastExprClass:
-        case Expr::CStyleCastExprClass:
-        case Expr::BoundsCastExprClass:
-          return IsCastExprInvertible(LValue, cast<CastExpr>(E));
-        default:
-          return false;
-      }
-    }
-
-    // Returns true if a unary operator is invertible with respect to LValue.
-    bool IsUnaryOperatorInvertible(Expr *LValue, UnaryOperator *E) {
-      Expr *SubExpr = E->getSubExpr()->IgnoreParens();
-      UnaryOperatorKind Op = E->getOpcode();
-
-      if (Op == UnaryOperatorKind::UO_AddrOf) {
-        // &*e1 is invertible with respect to LValue if e1 is invertible with
-        // respect to LValue.
-        if (UnaryOperator *UnarySubExpr = dyn_cast<UnaryOperator>(SubExpr)) {
-          if (UnarySubExpr->getOpcode() == UnaryOperatorKind::UO_Deref)
-            return IsInvertible(LValue, UnarySubExpr->getSubExpr());
-        }
-        // &e1[e2] is invertible with respect to LValue if e1 + e2 is invertible
-        // with respect to LValue.
-        else if (ArraySubscriptExpr *ArraySubExpr = dyn_cast<ArraySubscriptExpr>(SubExpr)) {
-          Expr *Base = ArraySubExpr->getBase();
-          Expr *Index = ArraySubExpr->getIdx();
-          BinaryOperator Sum(Context, Base, Index,
-                             BinaryOperatorKind::BO_Add,
-                             Base->getType(),
-                             Base->getValueKind(),
-                             Base->getObjectKind(),
-                             SourceLocation(),
-                             FPOptionsOverride());
-          return IsInvertible(LValue, &Sum);
-        }
-      }
-
-      // *&e1 is invertible with respect to LValue if e1 is invertible with
-      // respect to LValue.
-      if (Op == UnaryOperatorKind::UO_Deref) {
-        if (UnaryOperator *UnarySubExpr = dyn_cast<UnaryOperator>(SubExpr)) {
-          if (UnarySubExpr->getOpcode() == UnaryOperatorKind::UO_AddrOf)
-            return IsInvertible(LValue, UnarySubExpr->getSubExpr());
-        }
-      }
-
-      // ~e1, -e1, and +e1 are invertible with respect to LValue if e1 is
-      // invertible with respect to LValue.
-      if (Op == UnaryOperatorKind::UO_Not ||
-          Op == UnaryOperatorKind::UO_Minus ||
-          Op == UnaryOperatorKind::UO_Plus)
-        return IsInvertible(LValue, SubExpr);
-
-      return false;
-    }
-
-    // Returns true if a binary operator is invertible with respect to LValue.
-    bool IsBinaryOperatorInvertible(Expr *LValue, BinaryOperator *E) {
-      BinaryOperatorKind Op = E->getOpcode();
-      if (Op != BinaryOperatorKind::BO_Add &&
-          Op != BinaryOperatorKind::BO_Sub &&
-          Op != BinaryOperatorKind::BO_Xor)
-        return false;
-
-      Expr *LHS = E->getLHS();
-      Expr *RHS = E->getRHS();
-
-      // Addition and subtraction operations must be for checked pointer
-      // arithmetic or unsigned integer arithmetic.
-      if (Op == BinaryOperatorKind::BO_Add || Op == BinaryOperatorKind::BO_Sub) {
-        // The operation is checked pointer arithmetic if either the LHS
-        // or the RHS have checked pointer type.
-        bool IsCheckedPtrArithmetic = LHS->getType()->isCheckedPointerType() ||
-                                      RHS->getType()->isCheckedPointerType();
-        if (!IsCheckedPtrArithmetic) {
-          // The operation is unsigned integer arithmetic if both the LHS
-          // and the RHS have unsigned integer type.
-          bool IsUnsignedArithmetic = LHS->getType()->isUnsignedIntegerType() &&
-                                      RHS->getType()->isUnsignedIntegerType();
-          if (!IsUnsignedArithmetic)
-            return false;
-        }
-      }
-
-      // LValue must appear in exactly one subexpression of E and that
-      // subexpression must be invertible with respect to LValue.
-      std::pair<Expr *, Expr*> Pair = SplitByLValueCount(LValue, LHS, RHS);
-      if (!Pair.first)
-        return false;
-      Expr *E_LValue = Pair.first, *E_NotLValue = Pair.second;
-      if (!IsInvertible(LValue, E_LValue))
-        return false;
-
-      // The subexpression not containing LValue must be nonmodifying
-      // and cannot be or contain a pointer dereference, member
-      // reference, or indirect member reference.
-      if (!CheckIsNonModifying(E_NotLValue) ||
-          ReadsMemoryViaPointer(E_NotLValue, true))
-        return false;
-
-      return true;
-    }
-
-    // Returns true if a cast expression is invertible with respect to LValue.
-    // A cast expression (T1)e1 is invertible if T1 is a bit-preserving
-    // or widening cast and e1 is invertible.
-    bool IsCastExprInvertible(Expr *LValue, CastExpr *E) {
-      QualType T1 = E->getType();
-      QualType T2 = E->getSubExpr()->getType();
-      uint64_t Size1 = S.Context.getTypeSize(T1);
-      uint64_t Size2 = S.Context.getTypeSize(T2);
-
-      // If T1 is a smaller type than T2, then (T1)e1 is a narrowing cast.
-      if (Size1 < Size2)
-        return false;
-
-      switch (E->getCastKind()) {
-        // Bit-preserving casts
-        case CastKind::CK_BitCast:
-        case CastKind::CK_LValueBitCast:
-        case CastKind::CK_NoOp:
-        case CastKind::CK_ArrayToPointerDecay:
-        case CastKind::CK_FunctionToPointerDecay:
-        case CastKind::CK_NullToPointer:
-        // Widening casts
-        case CastKind::CK_BooleanToSignedIntegral:
-        case CastKind::CK_IntegralToFloating:
-          return IsInvertible(LValue, E->getSubExpr());
-        // Bounds casts may be invertible.
-        case CastKind::CK_DynamicPtrBounds:
-        case CastKind::CK_AssumePtrBounds: {
-          CHKCBindTemporaryExpr *Temp =
-            dyn_cast<CHKCBindTemporaryExpr>(E->getSubExpr());
-          assert(Temp);
-          return IsInvertible(LValue, Temp->getSubExpr());
-        }
-        // Potentially non-narrowing casts, depending on type sizes
-        case CastKind::CK_IntegralToPointer:
-        case CastKind::CK_PointerToIntegral:
-        case CastKind::CK_IntegralCast:
-          return Size1 >= Size2 && IsInvertible(LValue, E->getSubExpr());
-        // All other casts are considered narrowing.
-        default:
-          return false;
-      }
-    }
-
-    // Inverse repeatedly applies mathematical rules to the expression E to
-    // get the inverse of E with respect to the lvalue expression LValue and
-    // expression F. If rules cannot be applied to E, Inverse returns nullptr.
-    Expr *Inverse(Expr *LValue, Expr *F, Expr *E) {
-      if (!F)
-        return nullptr;
-
-      E = E->IgnoreParens();
-      Expr *RValueChild = ExprUtil::GetRValueCastChild(S, E);
-      if (RValueChild && EqualValue(S.Context, LValue, RValueChild, nullptr))
-        return F;
-
-      switch (E->getStmtClass()) {
-        case Expr::UnaryOperatorClass:
-          return UnaryOperatorInverse(LValue, F, cast<UnaryOperator>(E));
-        case Expr::BinaryOperatorClass:
-          return BinaryOperatorInverse(LValue, F, cast<BinaryOperator>(E));
-        case Expr::ImplicitCastExprClass:
-        case Expr::CStyleCastExprClass:
-        case Expr::BoundsCastExprClass:
-          return CastExprInverse(LValue, F, cast<CastExpr>(E));
-        default:
-          return nullptr;
-      }
-
-      return nullptr;
-    }
-
-    // Returns the inverse of a unary operator.
-    Expr *UnaryOperatorInverse(Expr *LValue, Expr *F, UnaryOperator *E) {
-      Expr *SubExpr = E->getSubExpr()->IgnoreParens();
-      UnaryOperatorKind Op = E->getOpcode();
-      
-      if (Op == UnaryOperatorKind::UO_AddrOf) {
-        // Inverse(f, &*e1) = Inverse(f, e1)
-        if (UnaryOperator *UnarySubExpr = dyn_cast<UnaryOperator>(SubExpr)) {
-          if (UnarySubExpr->getOpcode() == UnaryOperatorKind::UO_Deref)
-            return Inverse(LValue, F, UnarySubExpr->getSubExpr());
-        }
-        // Inverse(f, &e1[e2]) = Inverse(f, e1 + e2)
-        else if (ArraySubscriptExpr *ArraySubExpr = dyn_cast<ArraySubscriptExpr>(SubExpr)) {
-          Expr *Base = ArraySubExpr->getBase();
-          Expr *Index = ArraySubExpr->getIdx();
-          BinaryOperator Sum(Context, Base, Index,
-                             BinaryOperatorKind::BO_Add,
-                             Base->getType(),
-                             Base->getValueKind(),
-                             Base->getObjectKind(),
-                             SourceLocation(),
-                             FPOptionsOverride());
-          return Inverse(LValue, F, &Sum);
-        }
-      }
-
-      // Inverse(f, *&e1) = Inverse(f, e1)
-      if (Op == UnaryOperatorKind::UO_Deref) {
-        if (UnaryOperator *UnarySubExpr = dyn_cast<UnaryOperator>(SubExpr)) {
-          if (UnarySubExpr->getOpcode() == UnaryOperatorKind::UO_AddrOf)
-            return Inverse(LValue, F, UnarySubExpr->getSubExpr());
-        }
-      }
-
-      // Inverse(f, ~e1) = Inverse(~f, e1)
-      // Inverse(f, -e1) = Inverse(-f, e1)
-      // Inverse(f, +e1) = Inverse(+f, e1)
-      Expr *Child = ExprCreatorUtil::EnsureRValue(S, F);
-      Expr *F1 = UnaryOperator::Create(S.Context, Child, Op,
-                                       E->getType(),
-                                       E->getValueKind(),
-                                       E->getObjectKind(),
-                                       SourceLocation(),
-                                       E->canOverflow(),
-                                       FPOptionsOverride());
-      return Inverse(LValue, F1, SubExpr);
-    }
-
-    // Returns the inverse of a binary operator.
-    Expr *BinaryOperatorInverse(Expr *LValue, Expr *F, BinaryOperator *E) {
-      std::pair<Expr *, Expr*> Pair = SplitByLValueCount(LValue, E->getLHS(), E->getRHS());
-      if (!Pair.first)
-        return nullptr;
-
-      Expr *E_LValue = Pair.first, *E_NotLValue = Pair.second;
-      BinaryOperatorKind Op = E->getOpcode();
-      Expr *F1 = nullptr;
-
-      switch (Op) {
-        case BinaryOperatorKind::BO_Add:
-          // Inverse(f, e1 + e2) = Inverse(f - e_notlvalue, e_lvalue)
-          F1 = ExprCreatorUtil::CreateBinaryOperator(S, F, E_NotLValue, BinaryOperatorKind::BO_Sub);
-          break;
-        case BinaryOperatorKind::BO_Sub: {
-          if (E_LValue == E->getLHS())
-            // Inverse(f, e_lvalue - e_notlvalue) = Inverse(f + e_notlvalue, e_lvalue)
-            F1 = ExprCreatorUtil::CreateBinaryOperator(S, F, E_NotLValue, BinaryOperatorKind::BO_Add);
-          else
-            // Inverse(f, e_notlvalue - e_lvalue) => Inverse(e_notlvalue - f, e_lvalue)
-            F1 = ExprCreatorUtil::CreateBinaryOperator(S, E_NotLValue, F, BinaryOperatorKind::BO_Sub);
-          break;
-        }
-        case BinaryOperatorKind::BO_Xor:
-          // Inverse(f, e1 ^ e2) = Inverse(lvalue, f ^ e_notlvalue, e_lvalue)
-          F1 = ExprCreatorUtil::CreateBinaryOperator(S, F, E_NotLValue, BinaryOperatorKind::BO_Xor);
-          break;
-        default:
-          llvm_unreachable("unexpected binary operator kind");
-      }
-
-      return Inverse(LValue, F1, E_LValue);
-    }
-
-    // Returns the inverse of a cast expression.  If e1 has type T2,
-    // Inverse(f, (T1)e1) = Inverse((T2)f, e1) (assuming that (T1) is
-    // not a narrowing cast).
-    Expr *CastExprInverse(Expr *LValue, Expr *F, CastExpr *E) {
-      QualType T2 = E->getSubExpr()->getType();
-      switch (E->getStmtClass()) {
-        case Expr::ImplicitCastExprClass: {
-          Expr *F1 = CreateImplicitCast(T2, E->getCastKind(), F);
-          return Inverse(LValue, F1, E->getSubExpr());
-        }
-        case Expr::CStyleCastExprClass: {
-          Expr *F1 = CreateExplicitCast(T2, E->getCastKind(), F,
-                                        E->isBoundsSafeInterface());
-          return Inverse(LValue, F1, E->getSubExpr());
-        }
-        case Expr::BoundsCastExprClass: {
-          CHKCBindTemporaryExpr *Temp = dyn_cast<CHKCBindTemporaryExpr>(E->getSubExpr());
-          assert(Temp);
-          Expr *F1 = CreateExplicitCast(T2, CastKind::CK_BitCast, F,
-                                        E->isBoundsSafeInterface());
-          return Inverse(LValue, F1, Temp->getSubExpr());
-        }
-        default:
-          llvm_unreachable("unexpected cast kind");
-      }
       return nullptr;
     }
 
@@ -5801,7 +5391,7 @@ namespace {
         const AbstractSet *A = Pair.first;
         BoundsExpr *B = Pair.second;
         auto It = Context2.find(A);
-        if (It == Context2.end() || !EqualValue(Context, B, It->second, nullptr)) {
+        if (It == Context2.end() || !ExprUtil::EqualValue(Context, B, It->second, nullptr)) {
           Difference[A] = B;
         }
       }
@@ -5819,7 +5409,7 @@ namespace {
         auto It = Context2.find(Pair.first);
         if (It == Context2.end())
           return false;
-        if (!EqualValue(Context, Pair.second, It->second, nullptr))
+        if (!ExprUtil::EqualValue(Context, Pair.second, It->second, nullptr))
           return false;
       }
 
@@ -5904,7 +5494,7 @@ namespace {
         ExprSetTy F = *OuterList;
         for (auto InnerList = F.begin(); InnerList != F.end(); ++InnerList) {
           Expr *E1 = *InnerList;
-          if (EqualValue(S.Context, E, E1, nullptr)) {
+          if (ExprUtil::EqualValue(S.Context, E, E1, nullptr)) {
             ValuePreservingE = E1;
             return F;
           }
@@ -5977,71 +5567,6 @@ namespace {
           return false;
         }
       }
-    }
-
-    // Returns true if the expression e reads memory via a pointer.
-    // IncludeAllMemberExprs is used to modify the behavior to return true if e
-    // is or contains a pointer dereference, member reference, or indirect
-    // member reference (including e1.f which may not read memory via a
-    // pointer). Returns false if E is nullptr.
-    static bool ReadsMemoryViaPointer(Expr *E, bool IncludeAllMemberExprs = false) {
-      if (!E)
-        return false;
-
-      E = E->IgnoreParens();
-
-      switch (E->getStmtClass()) {
-        case Expr::UnaryOperatorClass: {
-          UnaryOperator *UO = cast<UnaryOperator>(E);
-          // *e reads memory via a pointer.
-          return UO->getOpcode() == UnaryOperatorKind::UO_Deref;
-        }
-        // e1[e2] is a synonym for *(e1 + e2), which reads memory via a pointer.
-        case Expr::ArraySubscriptExprClass:
-          return true;
-        case Expr::MemberExprClass: {
-          if (IncludeAllMemberExprs)
-            return true;
-
-          MemberExpr *ME = cast<MemberExpr>(E);
-          // e1->f reads memory via a pointer.
-          if (ME->isArrow())
-            return true;
-          // e1.f reads memory via a pointer if and only if e1 reads
-          // memory via a pointer.
-          else
-            return ReadsMemoryViaPointer(ME->getBase(), IncludeAllMemberExprs);
-        }
-        default: {
-          for (auto I = E->child_begin(); I != E->child_end(); ++I) {
-            if (Expr *SubExpr = dyn_cast<Expr>(*I)) {
-              if (ReadsMemoryViaPointer(SubExpr, IncludeAllMemberExprs))
-                return true;
-            }
-          }
-          return false;
-        }
-      }
-    }
-
-    // If LValue appears exactly once in Ei and does not appear in Ej,
-    // SplitByLValueCount returns the pair (Ei, Ej).  Otherwise, it returns
-    // an empty pair.
-    std::pair<Expr *, Expr *> SplitByLValueCount(Expr *LValue,
-                                                 Expr *E1, Expr *E2) {
-      std::pair<Expr *, Expr *> Pair;
-      unsigned int Count1 = LValueOccurrenceCount(S, LValue, E1);
-      unsigned int Count2 = LValueOccurrenceCount(S, LValue, E2);
-      if (Count1 == 1 && Count2 == 0) {
-        // LValue appears once in E1 and does not appear in E2.
-        Pair.first = E1;
-        Pair.second = E2;
-      } else if (Count2 == 1 && Count1 == 0) {
-        // LValue appears once in E2 and does not appear in E1.
-        Pair.first = E2;
-        Pair.second = E1;
-      }
-      return Pair;
     }
 
     // SynthesizeMembers modifies the set AbstractSets to include AbstractSets
@@ -6121,13 +5646,6 @@ namespace {
       }
 
       SynthesizeMembers(Base, M, CSS, AbstractSets);
-    }
-
-    // CheckIsNonModifying suppresses diagnostics while checking
-    // whether e is a non-modifying expression.
-    bool CheckIsNonModifying(Expr *E) {
-      return S.CheckIsNonModifying(E, Sema::NonModifyingContext::NMC_Unknown,
-                                   Sema::NonModifyingMessage::NMM_None);
     }
 
     BoundsExpr *CreateBoundsUnknown() {
@@ -6473,7 +5991,8 @@ namespace {
         };
 
         if (TargetTy != E->getType())
-          Base = CreateExplicitCast(TargetTy, CK_BitCast, Base, true);
+          Base = ExprCreatorUtil::CreateExplicitCast(S, TargetTy, CK_BitCast,
+                                                     Base, true);
       } else
         assert(Ty == E->getType());
 
@@ -6979,10 +6498,8 @@ Expr *Sema::MakeAssignmentImplicitCastExplicit(Expr *E) {
   if (isUsualUnaryConversion)
     return E;
 
-  PrepassInfo Info;
-  std::pair<ComparisonSet, ComparisonSet> EmptyFacts;
-  return CheckBoundsDeclarations(*this, Info, EmptyFacts).CreateExplicitCast(TargetTy, CK, SE,
-                                                             ICE->isBoundsSafeInterface());
+  return ExprCreatorUtil::CreateExplicitCast(*this, TargetTy, CK, SE,
+                                             ICE->isBoundsSafeInterface());
 }
 
 void Sema::CheckFunctionBodyBoundsDecls(FunctionDecl *FD, Stmt *Body) {
